Starting to add non-local Dirac ODE

+++ src/DiracODE/Adams_coefs.hpp
@@ -8,8 +8,8 @@ template <> struct AdamsCoefs<8> {
   static const int N = 8;
   const int AMa[N] = {-33953,   312874,  -1291214, 3146338,
                       -5033120, 5595358, -4604594, 4467094};
-  const double AMd = 1. / 3628800;
-  const int AMaa = 1070017.;
+  const double AMd = 1.0 / 3628800;
+  const double AMaa = 1070017.0;
   // Outward coefs:
   const int OIe[N][N] = {
       {-1338, 2940, -2940, 2450, -1470, 588, -140, 15},
@@ -21,15 +21,15 @@ template <> struct AdamsCoefs<8> {
       {140, -588, 1470, -2450, 2940, -2940, 1338, 105},
       {-960, 3920, -9408, 14700, -15680, 11760, -6720, 2283}};
   const int OIa[N] = {-105, 15, -5, 3, -3, 5, -15, 105};
-  const int OId = 840.;
+  const int OId = 840.0;
 };
 //------------------------------------------------------------------------------
 template <> struct AdamsCoefs<7> {
   // Adams coefs:
   static const int N = 7;
   const int AMa[N] = {1375, -11351, 41499, -88547, 123133, -121797, 139849};
-  const double AMd = 1. / 120960;
-  const int AMaa = 36799;
+  const double AMd = 1.0 / 120960;
+  const double AMaa = 36799.0;
   // Outward coefs:
   const int OIe[N][N] = {{-609, 1260, -1050, 700, -315, 84, -10},
                          {-140, -329, 700, -350, 140, -35, 4},
@@ -46,8 +46,8 @@ template <> struct AdamsCoefs<6> {
   // Adams coefs:
   static const int N = 6;
   const int AMa[N] = {-863, 6312, -20211, 37504, -46461, 65112};
-  const double AMd = 1. / 60480;
-  const int AMaa = 19087;
+  const double AMd = 1.0 / 60480;
+  const double AMaa = 19087.0;
   // Outward coefs:
   const int OIe[N][N] = {
       {-77, 150, -100, 50, -15, 2}, {-24, -35, 80, -30, 8, -1},
@@ -61,8 +61,8 @@ template <> struct AdamsCoefs<5> {
   // Adams coefs:
   static const int N = 5;
   const int AMa[N] = {27, -173, 482, -798, 1427};
-  const double AMd = 1. / 1440;
-  const int AMaa = 475;
+  const double AMd = 1.0 / 1440;
+  const double AMaa = 475.0;
   // Outward coefs:
   const int OIe[N][N] = {{-65, 120, -60, 20, -3},
                          {-30, -20, 60, -15, 2},
